requirements.txt
fastapi
uvicorn
pillow
pytesseract
transformers
torch
pytest
